Allow override of SelfPath
When creating the shell hooks, direnv injects executable path of itself
into the hook. If direnv is wrapped from a shell script to set
configuration variables, E.G. nixpkgs' `makeWrapper`, this causes it to
unwrap itself and ignore any set config.

direnv now checks the DIRENV_SELF_PATH environment variable to find the
location of the executable to call during hooks. A wrapper script should
pass its own $0 to this variable, so the hooks target that wrapper
instead of the underlying binary.